Restructured code
Split listener 'event' methods into partial class files
