+++ F1 Racing Hub/AppForm.cs
@@ -1,0 +1,11 @@
+namespace F1_Racing_Hub
+{
+    public partial class AppForm : Form
+    {
+        public AppForm()
+        {
+            listenerObject = new RacingHubListener();
+            InitializeComponent();
+        }
+    }
+}
