--- F1 Racing Hub/Form1.cs
@@ -1,11 +1,0 @@
-namespace F1_Racing_Hub
-{
-    public partial class Form1 : Form
-    {
-        public Form1()
-        {
-            listenerObject = new ListenerObject();
-            InitializeComponent();
-        }
-    }
-}
