+++ F1 Racing Hub/Utilities/SqlValueExtensions.cs
